refactor: harden pipeline streaming path
